backend/requirements.txt
fastapi
uvicorn
SQLAlchemy==2.0.*
pydantic==2.5.*
pydantic-settings
pre-commit==3.5.0
python-jose[cryptography]==3.3.*
requests==2.31.*
# psycopg2==2.9.*
psycopg2-binary ==2.9.5
passlib[bcrypt]==1.7.*
python-multipart
pycryptodome==3.19.1
openpyxl==3.1.2
pandas==2.1.4
openai==1.6.1
langchain==0.0.354
langchain-experimental==0.0.54
langchain-openai==0.0.8
scikit-image==0.22.0
tiktoken==0.5.2
pandas==2.1.4
lxml==5.1.0
tabulate==0.9.0
openpyxl
plotly==5.15.0
statsmodels==0.14.0
xlsxwriter==3.1.9
python-pptx==0.6.23
apscheduler==3.10.4
markdown-it-py==3.0.0
loguru==0.7.2
forex-python==1.8
jsonschema==4.21.1
darts
gpytorch
jax[cpu]
python-dotenv
tqdm
